# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS API Documentation",
        "description": "API documentation for LMS Smart Notebook project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/import/convert-to-html": {
            "post": {
                "tags": [
                    "Import"
                ],
                "summary": "Convert a Word file to HTML (with images) and store assets by note/student",
                "description": "Uploads a .doc/.docx file and note_id, converts it to HTML using LibreOffice, saves all images and HTML under structured folders, and returns the HTML.",
                "operationId": "389652dc6c7be5eceb459a75c5920fd7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "note_id"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Word (.doc/.docx) file to convert",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "note_id": {
                                        "description": "The ID of the note",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conversion successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "html": {
                                            "description": "HTML content of the converted Word file",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Conversion failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}